FIX : skip remove env dir if exit code 1

+++ run-simulation.sh
@@ -27,7 +27,10 @@ cmake . &> "$result_dir/log/cmake.log"
 make &> "$result_dir/log/make.log"
 
 GSL_RNG_SEED="$seed"  ./CLoTH_Gossip "$result_dir/" "$dijkstra_cache_filename" &> "$result_dir/log/cloth.log"
-cat "$result_dir/output.log"
-echo "seed=$seed" >> "$result_dir/cloth_input.txt"
-
-rm -Rf "$environment_dir"
+if [ $? -eq 0 ]; then
+  echo "seed=$seed" >> "$result_dir/cloth_input.txt"
+  rm -Rf "$environment_dir"
+else
+  echo "error $?"
+  echo "seed=$seed" >> "$result_dir/cloth_input.txt"
+fi
